requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
analytics-python==1.4.0
anyio==3.5.0
asgiref==3.5.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
backoff==1.10.0
bcrypt==3.2.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
cycler==0.11.0
fastapi==0.73.0
ffmpy==0.3.0
flatbuffers==2.0
fonttools==4.29.1
frozenlist==1.3.0
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gradio==2.7.5.2
grpcio==1.43.0
h11==0.13.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.10.1
Jinja2==3.0.3
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
Markdown==3.3.6
markdown2==2.4.2
MarkupSafe==2.0.1
matplotlib==3.5.1
monotonic==1.6
multidict==6.0.2
numpy==1.22.2
oauthlib==3.2.0
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.0
paramiko==2.9.2
Pillow==9.0.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.1
pydantic==1.9.0
pydub==0.25.1
PyNaCl==1.5.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.4
Werkzeug==2.0.2
wrapt==1.13.3
yarl==1.7.2
zipp==3.7.0
